refactor(ai-tracing): move log level mapping to base exporter
Eliminates duplicated log level mapping code across all exporters:
- Add resolveLogLevel() method to BaseExporter
- Accept both LogLevel enum and string literals in BaseExporterConfig
- Remove duplicate mapping logic from LangfuseExporter, LangSmithExporter, BraintrustExporter
- Remove LogLevel import from observability packages (no longer needed)

This centralizes the string-to-enum conversion in one place,
making the code DRYer and easier to maintain.
